requirements.txt
streamlit
python-dotenv
feedparser
openai